refactor(DAO-2135): move banner components into subfolder and rewrite root barrel
Move BtcVaultBanners, BtcVaultEligibilityAndDepositCard,
BtcVaultPrototypeBannerContent, EligibilityBannerContent,
PauseBannerContent, DepositWindowBanner, DepositWindowSection, and
btcVaultBannerGradients into banners/ subfolder. Rewrite root
components/index.ts to re-export from new subfolder barrels.
