+++ src/app/btc-vault/components/banners/BtcVaultPrototypeBannerContent.tsx
