+++ src/app/btc-vault/components/banners/EligibilityBannerContent.test.tsx
