+++ src/app/btc-vault/components/banners/EligibilityBannerContent.tsx
