+++ src/app/btc-vault/components/banners/btcVaultBannerGradients.ts
