+++ src/app/btc-vault/components/banners/index.ts
@@ -1,0 +1,7 @@
+export { BtcVaultBanners } from './BtcVaultBanners'
+export {
+  EligibilityBannerContent,
+  type EligibilityBannerContentProps,
+  type EligibilityBannerVariant,
+} from './EligibilityBannerContent'
+export { PauseBannerContent, type PauseBannerContentProps } from './PauseBannerContent'
