+++ src/app/btc-vault/components/index.ts
@@ -1,8 +1,8 @@
-export { BtcVaultActions } from './BtcVaultActions'
-export { BtcVaultDashboard } from './BtcVaultDashboard'
 export {
   EligibilityBannerContent,
   type EligibilityBannerContentProps,
   type EligibilityBannerVariant,
-} from './EligibilityBannerContent'
-export { PauseBannerContent, type PauseBannerContentProps } from './PauseBannerContent'
+  PauseBannerContent,
+  type PauseBannerContentProps,
+} from './banners'
+export { BtcVaultActions, BtcVaultDashboard } from './dashboard'
